Modularised the SASS and JS structure
