+++ src/js/app.js
@@ -2,6 +2,7 @@
 
 const React = require('react');
 const ReactDOM = require('react-dom');
+const Episodes = require('./episodes');
 
 console.log(React);
 
@@ -33,234 +34,8 @@ const json = [
 	}
 ];
 
-const Svg = React.createClass({
-
-	path: {
-		item: 'M10 16.5l6-4.5-6-4.5v9zm2-14.5c-5.5 0-10 4.5-10 10s4.5 10 10 10 10-4.5 10-10-4.5-10-10-10zm0 19c-5 0-9-4-9-9s4-9 9-9 9 4 9 9-4 9-9 9z',
-		play: 'M8 5v14l11-7z',
-		pause: 'M6 19h4v-14h-4v14zm8-14v14h4v-14h-4z',
-		skip: 'M4 18l8.5-6-8.5-6v12zm9-12v12l8.5-6-8.5-6z'
-	},
-
-	getIconClass() {
-
-		return `${this.props.component}__icon icon--${this.props.icon}`;
-
-	},
-
-	render() {
-
-		const className = this.getIconClass();
-		const path = this.path[this.props.icon];
-
-		return (
-			<svg className={className} height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
-				<path d={path}/>
-			</svg>
-		);
-
-	}
-
-});
-
-const Player = React.createClass({
-
-	getInitialState() {
-
-		return {
-			track: this.loadTrack(),
-			phase: 'play'
-		};
-
-	},
-
-	componentDidMount() {
-
-		const player = ReactDOM.findDOMNode(this);
-
-		this.positionPing(player);
-		this.activatePlayer(player);
-		this.trackPhase('play');
-
-	},
-
-	componentWillUnmount() {
-
-		// Stops auto from continuing to play when player is no longer relevant.
-		this.trackPhase('stop');
-
-	},
-
-	loadTrack() {
-
-		return new Audio(this.props.json.mp3);
-
-	},
-
-	trackPhase(phase) {
-
-		const action = phase === 'play' ? 'play' : 'pause';
-
-		this.state.track[action]();
-		if (phase === 'stop') this.state.track.src = 'data:audio/wav;base64,UklGRiQAAABXQVZFZm10IBAAAAABAAEAVFYAAFRWAAABAAgAZGF0YQAAAAA=';
-		this.setState({phase: phase});
-
-	},
-
-	togglePhase() {
-
-		const phase = this.state.phase === 'play' ? 'pause' : 'play';
-
-		this.trackPhase(phase);
-
-	},
-
-	positionPing(player) {
-
-		const background = player.getElementsByClassName('player__background')[0];
-		const bounds = background.getBoundingClientRect();
-		const cursor = this.props.cursor;
-		const x = cursor.left;
-		const y = cursor.top - bounds.top;
-
-		background.style.left = `${x}px`;
-		background.style.top = `${y}px`;
-
-	},
-
-	activatePlayer(player) {
-
-		setTimeout(() => player.classList.add('player--active'), 0);
-
-	},
-
-	render() {
-
-		return (
-			<div className="player">
-				<div className="player__background"></div>
-				<h3 className="player__title">{this.props.json.title}</h3>
-				<ul className="player__controls">
-					<li className="player__control">
-						<button className="player__button player__button--back" type="button" name="skip forward">
-							<Svg component={'player'} icon={'skip'}/>
-						</button>
-					</li>
-					<li className="player__control">
-						{
-
-							(() => {
-
-								const phase = this.state.phase;
-								const className = `player__button player__button--${phase}`;
-
-								return (
-									<button onClick={this.togglePhase} className={className} type="button" name={phase}>
-										<Svg component={'player'} icon={phase}/>
-									</button>
-								);
-
-							})()
-
-						}
-					</li>
-					<li className="player__control">
-						<button className="player__button player__button--forward" type="button" name="skip back">
-							<Svg component={'player'} icon={'skip'}/>
-						</button>
-					</li>
-				</ul>
-				<div className="player__time">
-					<progress className="player__progress" max="100" value="80"></progress>
-					<span className="player__elapsed">12.45</span>
-					<span className="player__total">34.27</span>
-				</div>
-			</div>
-		);
-
-	}
-
-});
-
-const App = React.createClass({
-
-	getInitialState() {
-
-		return {
-			active: null,
-			cursor: {left: null, top: null}
-		};
-
-	},
-
-	getEpisodeState(current) {
-
-		const active = this.state.active;
-		const className = 'episode';
-		const modifier = active === null ? '' :
-                         current === active ? `${className}--active` :
-                         current > active ? `${className}--dormantBefore` :
-                         `${className}--dormantAfter`;
-
-		return `${className} ${modifier}`;
-
-	},
-
-	activateEpisode(current, e) {
-
-		this.setState({
-			active: current,
-			cursor: this.cursorLocation(e)
-		});
-
-	},
-
-	cursorLocation(e) {
-
-		const left = e.clientX;
-		const top = e.clientY;
-
-		return {left, top};
-
-	},
-
-	render() {
-
-		return (
-
-			<ul className="episodes__list">
-				{
-
-					this.props.json.map((episode, current) => {
-
-						const state = this.getEpisodeState(current);
-
-						return (
-							<li className={state} key={current} onClick={this.activateEpisode.bind(this, current)}>
-								<a className="episode__link">
-									<Svg component={'episode'} icon={'item'}/>
-									<h2 className="episode__title">{episode.title}</h2>
-									<span className="episode__date">12 February 2016</span>
-									<p className="episode__description">Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
-								</a>
-								{this.state.active === current ? <Player json={this.props.json[current]} cursor={this.state.cursor}/> : ''}
-							</li>
-
-						);
-
-					})
-
-				}
-			</ul>
-
-		);
-
-	}
-
-});
-
 ReactDOM.render(
-	<App json={json}/>,
+	<Episodes json={json}/>,
 	document.getElementById('episodes')
 );
