+++ src/sass/_app.scss
@@ -8,223 +8,3 @@
     // max-height: 667px;
     // max-width: 375px;
 }
-
-.episodes {
-
-    &__list {
-        list-style: none;
-        margin: 0;
-        padding: 0;
-    }
-}
-
-.episode {
-    padding: space(small) space();
-    position: relative;
-    transition: transform speed() ease();
-
-    &--dormantBefore {
-        transform: translateY(30px);
-    }
-
-    &--dormantActive {
-
-    }
-
-    &--dormantAfter {
-        transform: translateY(-30px);
-    }
-
-    &__icon {
-        fill: color(gray, dark);
-        height: space();
-        left: space();
-        position: absolute;
-        top: space(small);
-        width: space();
-    }
-
-    &__title,
-    &__date,
-    &__description {
-        @include ellipsis;
-        padding-left: space() * 1.5;
-    }
-
-    &__title {
-        color: color(gray, dark);
-        font: {
-            size: 14px;
-            weight: 300;
-        }
-        margin: 0 0 -3px;
-    }
-
-    &__date {
-        color: color(gray);
-        font: {
-            size: 9px;
-            weight: 700;
-        }
-        text-transform: uppercase;
-    }
-
-    &__description {
-        color: color(gray, dark);
-        font: {
-            size: 11px;
-            weight: 300;
-        }
-        margin: 0;
-    }
-}
-
-.player {
-    color: white;
-    left: 0;
-    overflow: hidden;
-    padding: space() space() space(small);
-    position: absolute;
-    right: 0;
-    top: 50%;
-    transform: translateY(-50%) translateZ(0);
-    z-index: zIndex(player);
-
-    > *:not(.player__background) {
-        opacity: 0;
-        transform: translateX(space(small));
-        transition: speed(fast) ease();
-        transition-property: opacity transform;
-    }
-
-    &.player--active {
-
-        .player__background {
-            transform: scale(1);
-        }
-
-        > * {
-            opacity: 1;
-            transform: translateX(0);
-
-            // Do not target .player__background.
-            @for $i from 2 through 4 {
-
-                &:nth-child(#{$i}) {
-                    transition-delay: delay(fast) * ($i - 1);
-                }
-            }
-        }
-    }
-
-    &.player--dormant {
-        opacity: 0;
-
-        // > *:not(.player__background) {
-        //     opacity: 0;
-        // }
-    }
-
-    &__background {
-        height: 100%;
-        left: 0;
-        position: absolute;
-        top: 0;
-        transform: scale(0);
-        transform-origin: left top;
-        transition: transform speed() ease();
-        width: 100%;
-
-        &:before {
-            background: color(teal);
-            border-radius: 50%;
-            content: '';
-            display: block;
-            height: 0;
-            padding-top: 216%;
-            left: 0;
-            position: absolute;
-            top: 0;
-            transform: translate(-50%, -50%);
-            width: 216%;
-        }
-    }
-
-    &__title {
-        @include ellipsis;
-        font: {
-            size: 14px;
-            weight: 300;
-        }
-        margin: 0 0 space(small);
-    }
-
-    &__controls {
-        display: flex;
-        justify-content: space-between;
-        list-style: none;
-        margin: 0;
-        padding: 0;
-    }
-
-    &__button {
-        @include resetButton;
-        cursor: pointer;
-
-        &--back {
-            transform: rotate(180deg);
-        }
-    }
-
-    &__icon {
-        fill: white;
-        height: space();
-        width: space();
-    }
-
-    &__time {
-        display: flex;
-        flex-wrap: wrap;
-        justify-content: space-between;
-    }
-
-    &__progress {
-        appearance: none;
-        background: color(teal, light);
-        border: 0;
-        flex-basis: 100%;
-        height: 2px;
-        margin: space(tiny) 0;
-
-        // Background color.
-        & { // Firefox.
-            background: color(teal, light);
-        }
-
-        &::-webkit-progress-bar { // Chrome.
-            background: color(teal, light);
-        }
-
-        // Forground color.
-        & { // IE.
-            color: white;
-        }
-
-        &::-moz-progress-bar { // Firefox.
-            background: white;
-        }
-
-        &::-webkit-progress-value { // Chrome.
-            background: white;
-        }
-    }
-
-    &__elapsed,
-    &__total {
-        font-size: 11px;
-    }
-
-    &__total {
-        color: color(teal, light);
-    }
-}
