--- src/sass/_feed.scss
@@ -1,11 +1,0 @@
-.podcast__feed {
-
-}
-
-.podcast__episode {
-    color: skyblue;
-
-    &--active {
-        color: darkblue;
-    }
-}
