+++ src/sass/_player.scss
@@ -1,0 +1,171 @@
+.player {
+    color: white;
+    left: 0;
+    overflow: hidden;
+    padding: space() space() space(small);
+    position: absolute;
+    right: 0;
+    top: 50%;
+    transform: translateY(-50%) translateZ(0);
+    z-index: zIndex(player);
+
+    > *:not(.player__background) {
+        opacity: 0;
+        transform: translateX(space(small));
+        transition: speed(fast) ease();
+        transition-property: opacity transform;
+    }
+
+    &.player--active {
+
+        .player__background {
+            transform: scale(1);
+        }
+
+        > * {
+            opacity: 1;
+            transform: translateX(0);
+
+            // 2++ = Do not target .player__background.
+            @for $i from 2 through 4 {
+
+                &:nth-child(#{$i}) {
+                    transition-delay: delay(fast) * ($i - 1);
+                }
+            }
+        }
+    }
+
+    &.player--fixed {
+        position: fixed;
+        transform: none;
+
+        &Bottom {
+            bottom: 0;
+        }
+
+        &Top {
+            top: 0;
+        }
+    }
+
+    &__background {
+        height: 100%;
+        left: 0;
+        position: absolute;
+        top: 0;
+        transform: scale(0);
+        transform-origin: left top;
+        transition: transform speed() ease();
+        width: 100%;
+
+        &:before {
+            background: color(teal);
+            border-radius: 50%;
+            content: '';
+            display: block;
+            height: 0;
+            padding-top: 216%;
+            left: 0;
+            position: absolute;
+            top: 0;
+            transform: translate(-50%, -50%);
+            width: 216%;
+        }
+    }
+
+    &__title {
+        @include ellipsis;
+        font: {
+            size: 14px;
+            weight: 300;
+        }
+        margin: 0 0 space(small);
+    }
+
+    &__controls {
+        display: flex;
+        justify-content: space-between;
+        list-style: none;
+        margin: 0;
+        padding: 0;
+    }
+
+    &__button {
+        @include resetButton;
+        cursor: pointer;
+        outline: none;
+
+        &--back {
+            transform: rotate(180deg);
+        }
+
+        &:hover,
+        &:focus {
+
+            .player__icon {
+                fill: color(teal, dark);
+            }
+        }
+    }
+
+    &__icon {
+        fill: white;
+        height: space();
+        transition: fill speed(fast) ease();
+        width: space();
+    }
+
+    &__time {
+        display: flex;
+        flex-wrap: wrap;
+        justify-content: space-between;
+    }
+
+    &__progress {
+        appearance: none;
+        background: color(teal, light);
+        border: 0;
+        flex-basis: 100%;
+        height: 2px;
+        margin: space(tiny) 0;
+
+        // Background color.
+
+        // Firefox.
+        & {
+            background: color(teal, light);
+        }
+
+        // Chrome.
+        &::-webkit-progress-bar {
+            background: color(teal, light);
+        }
+
+        // Forground color.
+
+        // IE.
+        & {
+            color: white;
+        }
+
+        // Firefox.
+        &::-moz-progress-bar {
+            background: white;
+        }
+
+        // Chrome.
+        &::-webkit-progress-value {
+            background: white;
+        }
+    }
+
+    &__elapsed,
+    &__total {
+        font-size: 11px;
+    }
+
+    &__total {
+        color: color(teal, light);
+    }
+}
