+++ src/sass/style.scss
@@ -1,4 +1,5 @@
 @import 'var';
 @import 'base';
 @import 'app';
-@import 'feed';
+@import 'episodes';
+@import 'player';
